refactor: move components and types outside app
