+++ client/app/page.tsx
@@ -1,14 +1,10 @@
-import Image from "next/image";
-import DraftCard from "./components/DraftCard";
+import DraftCard from "@/components/DraftCard";
 import blogData from "@/mock-data/data";
 
 export default function Home() {
   return (
-    <div className="">
-      KZBlogs
-      <div>
-        <DraftCard blogs={blogData} />
-      </div>
+    <div className="w-[15vw] m-6">
+      <DraftCard blogs={blogData} />
     </div>
   );
 }
