+++ client/types/index.ts
@@ -1,13 +1,13 @@
 export interface Comments{
-    author_id:string
+    authorId:string
     content:string
 }
 
 export interface Blog{
     id:number
     name:string
-    publish_status:string
-    last_edited:string
+    publishStatus:string
+    lastEdited:string
     author:string
     likes:number
     comments ?:Comments[]
@@ -20,7 +20,7 @@ export interface User{
     username:string
     avatar:string
     bookmarks:Blog[]
-    google_id:string
-    is_admin:boolean
+    googleId:string
+    isAdmin:boolean
 }
